refactor: move ModularIsEqual GPU chip and FFI into algebra cuda module
Per AGENTS.md, CUDA prover code lives under cuda/ — mirrors the
mod-builder crate structure (cuda/mod.rs + cuda/cuda_abi.rs).
