+++ extensions/algebra/circuit/src/extension/hybrid.rs
@@ -34,6 +34,7 @@ use openvm_rv32im_circuit::Rv32ImGpuProverExt;
 use openvm_stark_backend::prover::AirProvingContext;
 
 use crate::{
+    cuda::ModularIsEqualChipGpu,
     fp2_chip::{get_fp2_addsub_chip, get_fp2_muldiv_chip, Fp2Air, Fp2Chip},
     modular_chip::*,
     AlgebraRecord, Fp2Extension, ModularExtension, Rv32ModularConfig, Rv32ModularWithFp2Config,
@@ -473,162 +474,5 @@ impl VmBuilder<E> for Rv32ModularWithFp2HybridBuilder {
             inventory,
         )?;
         Ok(chip_complex)
-    }
-}
-
-// ---------------------------------------------------------------------------
-// GPU tracegen for ModularIsEqual (dedicated kernel; see cuda/src/modular_is_eq.cu)
-// ---------------------------------------------------------------------------
-
-mod is_eq_cuda_abi {
-    #![allow(clippy::missing_safety_doc, clippy::too_many_arguments)]
-    use openvm_cuda_backend::prelude::F;
-    use openvm_cuda_common::{d_buffer::DeviceBuffer, error::CudaError, stream::cudaStream_t};
-
-    macro_rules! declare_is_eq_launcher {
-        ($name:ident) => {
-            extern "C" {
-                fn $name(
-                    d_trace: *mut F,
-                    height: usize,
-                    rows_used: usize,
-                    d_records: *const u8,
-                    rec_stride: usize,
-                    rec_core_offset: usize,
-                    d_modulus_limbs: *const u8,
-                    d_range_checker: *mut u32,
-                    rc_bins: usize,
-                    d_bitwise_lookup: *mut u32,
-                    bitwise_num_bits: usize,
-                    pointer_max_bits: u32,
-                    timestamp_max_bits: u32,
-                    stream: cudaStream_t,
-                ) -> i32;
-            }
-        };
-    }
-    declare_is_eq_launcher!(_modular_is_eq_tracegen_l8);
-    declare_is_eq_launcher!(_modular_is_eq_tracegen_l12);
-
-    pub unsafe fn tracegen(
-        d_trace: &DeviceBuffer<F>,
-        height: usize,
-        rows_used: usize,
-        d_records: &DeviceBuffer<u8>,
-        rec_stride: usize,
-        rec_core_offset: usize,
-        d_modulus_limbs: &DeviceBuffer<u8>,
-        d_range_checker: &DeviceBuffer<F>,
-        d_bitwise_lookup: &DeviceBuffer<F>,
-        num_lanes: usize,
-        pointer_max_bits: u32,
-        timestamp_max_bits: u32,
-        stream: cudaStream_t,
-    ) -> Result<(), CudaError> {
-        let launcher = match num_lanes {
-            8 => _modular_is_eq_tracegen_l8,
-            12 => _modular_is_eq_tracegen_l12,
-            _ => panic!("unsupported ModularIsEqual num_lanes {num_lanes}"),
-        };
-        CudaError::from_result(launcher(
-            d_trace.as_mut_ptr(),
-            height,
-            rows_used,
-            d_records.as_ptr(),
-            rec_stride,
-            rec_core_offset,
-            d_modulus_limbs.as_ptr(),
-            d_range_checker.as_mut_ptr() as *mut u32,
-            d_range_checker.len(),
-            d_bitwise_lookup.as_mut_ptr() as *mut u32,
-            8,
-            pointer_max_bits,
-            timestamp_max_bits,
-            stream,
-        ))
-    }
-}
-
-pub struct ModularIsEqualChipGpu<
-    const NUM_LANES: usize,
-    const LANE_SIZE: usize,
-    const TOTAL_LIMBS: usize,
-> {
-    range_checker: Arc<VariableRangeCheckerChipGPU>,
-    bitwise_lookup: Arc<BitwiseOperationLookupChipGPU<8>>,
-    d_modulus: openvm_cuda_common::d_buffer::DeviceBuffer<u8>,
-    pointer_max_bits: u32,
-    timestamp_max_bits: u32,
-}
-
-impl<const NUM_LANES: usize, const LANE_SIZE: usize, const TOTAL_LIMBS: usize>
-    ModularIsEqualChipGpu<NUM_LANES, LANE_SIZE, TOTAL_LIMBS>
-{
-    pub fn new(
-        modulus_limbs: [u8; TOTAL_LIMBS],
-        byte_ptr_max_bits: usize,
-        timestamp_max_bits: usize,
-        range_checker: Arc<VariableRangeCheckerChipGPU>,
-        bitwise_lookup: Arc<BitwiseOperationLookupChipGPU<8>>,
-    ) -> Self {
-        use openvm_cuda_common::copy::MemCopyH2D;
-        let d_modulus = modulus_limbs
-            .as_slice()
-            .to_device_on(&range_checker.device_ctx)
-            .unwrap();
-        Self {
-            range_checker,
-            bitwise_lookup,
-            d_modulus,
-            pointer_max_bits: byte_ptr_max_bits as u32,
-            timestamp_max_bits: timestamp_max_bits as u32,
-        }
-    }
-}
-
-impl<const NUM_LANES: usize, const LANE_SIZE: usize, const TOTAL_LIMBS: usize>
-    Chip<DenseRecordArena, GpuBackend>
-    for ModularIsEqualChipGpu<NUM_LANES, LANE_SIZE, TOTAL_LIMBS>
-{
-    fn generate_proving_ctx(&self, arena: DenseRecordArena) -> AirProvingContext<GpuBackend> {
-        use openvm_cuda_common::copy::MemCopyH2D;
-        let record_stride = size_of::<(
-            Rv32IsEqualModAdapterRecord<2, NUM_LANES, LANE_SIZE, TOTAL_LIMBS>,
-            ModularIsEqualRecord<TOTAL_LIMBS>,
-        )>();
-        let rec_core_offset =
-            size_of::<Rv32IsEqualModAdapterRecord<2, NUM_LANES, LANE_SIZE, TOTAL_LIMBS>>();
-        let records = arena.allocated();
-        if records.is_empty() {
-            return AirProvingContext::simple_no_pis(DeviceMatrix::dummy());
-        }
-        debug_assert_eq!(records.len() % record_stride, 0);
-        let rows_used = records.len() / record_stride;
-        let height = rows_used.next_power_of_two();
-        let width = Rv32IsEqualModAdapterCols::<F, 2, NUM_LANES, LANE_SIZE>::width()
-            + ModularIsEqualCoreCols::<F, TOTAL_LIMBS>::width();
-
-        let device_ctx = &self.range_checker.device_ctx;
-        let d_records = records.to_device_on(device_ctx).unwrap();
-        let d_trace = DeviceMatrix::<F>::with_capacity_on(height, width, device_ctx);
-        unsafe {
-            is_eq_cuda_abi::tracegen(
-                d_trace.buffer(),
-                height,
-                rows_used,
-                &d_records,
-                record_stride,
-                rec_core_offset,
-                &self.d_modulus,
-                &self.range_checker.count,
-                &self.bitwise_lookup.count,
-                NUM_LANES,
-                self.pointer_max_bits,
-                self.timestamp_max_bits,
-                device_ctx.stream.as_raw(),
-            )
-            .unwrap();
-        }
-        AirProvingContext::simple_no_pis(d_trace)
     }
 }
